requirements.txt
cachetools==3.1.1
Click==7.0
Flask==1.1.1
google-api-python-client==1.7.11
google-auth==1.9.0
google-auth-httplib2==0.0.3
httplib2==0.14.0
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.1
MarkupSafe==1.1.1
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyDrive==1.3.1
python-docx==0.8.10
PyYAML==5.2
rsa==4.0
six==1.13.0
uritemplate==3.0.0
Werkzeug==0.16.0
